JobRunnerPipeline: get json_decode() error
Some jobs have a status exit of 0 but are still logged as being in
error because the result is not an array.  I suspect that the job yields
an invalid JSON somehow.

Capture the PHP json last error number and message whenever
json_decode() fails (it returns null) and append them to the error
message.

Change-Id: I6b308b54ec8dfd7edfe45e605ebcafcd7cf7c935